requirements.txt
streamlit>=1.28
anthropic>=0.40
PyMuPDF>=1.24
fpdf2>=2.8
Pillow>=10.0
python-dotenv>=1.0
